Fix 8 bugs found during audit: LR controller, learnable weights, numerics
Critical (1-3):
- Fix encoder param group name casing (encoder_A -> encoder_a) so
  LR controller scales are actually applied to the optimizer
- Fix double gate computation in MetricBasedLR.update() that caused
  plateau counters to increment twice per epoch
- Add loss_fn.parameters() to optimizer when learnable_weights enabled

Moderate (4-8):
- Respect deterministic mode: skip cudnn.benchmark when active
- Pass weight_decay to RiemannianAdam (was silently omitted)
- Implement max_norm clamping in log_map_zero() (parameter was ignored);
  default is now curvature-aware (ball_radius - 1e-5)
- Use scheduler.get_last_lr() in LR controller so scales compose with
  cosine annealing instead of overriding it
- Use dtype=float64 for learnable log_sigma parameters to match codebase

+++ src/losses/combined.py
@@ -213,34 +213,34 @@ class CombinedLoss(nn.Module):
         # RichHierarchy sub-components
         if self.rich_hierarchy is not None:
             self.log_sigma_hierarchy = nn.Parameter(
-                torch.tensor(weight_to_log_sigma(self.rich_hierarchy_weights['hierarchy']))
+                torch.tensor(weight_to_log_sigma(self.rich_hierarchy_weights['hierarchy']), dtype=torch.float64)
             )
             self.log_sigma_coverage = nn.Parameter(
-                torch.tensor(weight_to_log_sigma(self.rich_hierarchy_weights['coverage']))
+                torch.tensor(weight_to_log_sigma(self.rich_hierarchy_weights['coverage']), dtype=torch.float64)
             )
             self.log_sigma_separation = nn.Parameter(
-                torch.tensor(weight_to_log_sigma(self.rich_hierarchy_weights['separation']))
+                torch.tensor(weight_to_log_sigma(self.rich_hierarchy_weights['separation']), dtype=torch.float64)
             )
 
         # Other losses
         if self.radial_loss is not None:
             self.log_sigma_radial = nn.Parameter(
-                torch.tensor(weight_to_log_sigma(self.radial_weight))
+                torch.tensor(weight_to_log_sigma(self.radial_weight), dtype=torch.float64)
             )
 
         if self.geodesic_loss is not None:
             self.log_sigma_geodesic = nn.Parameter(
-                torch.tensor(weight_to_log_sigma(self.geodesic_weight))
+                torch.tensor(weight_to_log_sigma(self.geodesic_weight), dtype=torch.float64)
             )
 
         if self.rank_loss is not None:
             self.log_sigma_rank = nn.Parameter(
-                torch.tensor(weight_to_log_sigma(self.rank_weight))
+                torch.tensor(weight_to_log_sigma(self.rank_weight), dtype=torch.float64)
             )
 
         if self.monotonic_loss is not None:
             self.log_sigma_monotonic = nn.Parameter(
-                torch.tensor(weight_to_log_sigma(self.monotonic_weight))
+                torch.tensor(weight_to_log_sigma(self.monotonic_weight), dtype=torch.float64)
             )
 
     def _uncertainty_weight(self, log_sigma: nn.Parameter) -> torch.Tensor:
